Add localstorage function
Created a save to local storage function then grabs that data on page load.
Changed the click event for going to project page. Instead of deleting project
elements and saving them in an array, simply grab objects from projectArray
and use that to create elements that will populate the projects page.
This accomodates both localstorage data and when back button is used in-page.
Strange undefined objects are present in array if accessed from localstorage,
and are not visible in the Applications tab of the devtools. Array.length is
more than actual objects visible. Presumably an artifact when data is set
to local storage by saveToLocalStorage function. If-block that looks for
undefined object then breaks out of the function introduced to bypass error.