Fix logic of local storage function
The problem was in fact starting from the initialize app
function. Once the saved data is assigned to the projects
and tasks array, the append functions for both container
and arrays were called which essentially doubled the number
of elements, and created undefined values. There was also no
check for when the array is empty, which created an empty
div in the "today" container. Further, while the projects
array displayed on the web page because of the gotoprojects
function, the "today" tasks didn't show up because the
todoListArray was not being looped through in the initialize
app function.